Transverse Mercator: Remove latitude of origin parameter
This is too buggy to include right now. Might revisit in the future
but the Karney/Krueger equations are not derived with a latitude of
origin. In addition Proj.4 seems to have some bugs with a latitude
of origin with a spherical projection.

All of the tests appear to pass now. :-D